Switch to pure Scala hpack
