+++ build.sbt
@@ -550,13 +550,17 @@ lazy val emberCore = libraryCrossProject("ember-core", CrossType.Full)
             .exclude[DirectMissingMethodProblem](
               "org.http4s.ember.core.Parser#MessageP.fromProduct"
             ),
+          ProblemFilters.exclude[MissingTypesProblem]("org.http4s.ember.core.h2.Hpack$"),
+          ProblemFilters.exclude[IncompatibleTemplateDefProblem](
+            "org.http4s.ember.core.h2.HpackPlatform"
+          ),
         )
       else Seq.empty
     },
   )
   .jsSettings(
     jsVersionIntroduced("0.23.5"),
-    mimaBinaryIssueFilters := {
+    mimaBinaryIssueFilters ++= {
       if (tlIsScala3.value)
         Seq(
           ProblemFilters.exclude[IncompatibleTemplateDefProblem](
@@ -565,20 +569,17 @@ lazy val emberCore = libraryCrossProject("ember-core", CrossType.Full)
           ProblemFilters.exclude[IncompatibleTemplateDefProblem](
             "org.http4s.ember.core.h2.facade.Decompressor"
           ),
+          ProblemFilters.exclude[Problem]("org.http4s.ember.core.h2.facade.*"),
         )
       else
         Seq.empty
     },
   )
   .jvmSettings(
-    libraryDependencies += "com.twitter" % "hpack" % "1.0.2"
+    libraryDependencies += twitterHpack
   )
-  .jsEnablePlugins(ScalaJSBundlerPlugin)
   .jsSettings(
-    Compile / npmDependencies += "hpack.js" -> "2.1.6",
-    useYarn := true,
-    yarnExtraArgs += "--frozen-lockfile",
-    Test / scalaJSLinkerConfig ~= (_.withModuleKind(ModuleKind.CommonJSModule)),
+    libraryDependencies += hpack.value
   )
   .dependsOn(core, testing % "test->test")
 
@@ -634,13 +635,10 @@ lazy val emberServer = libraryCrossProject("ember-server")
       jnrUnixSocket % Test, // Necessary for jdk < 16
     )
   )
-  .jsEnablePlugins(ScalaJSBundlerPlugin)
   .jsSettings(
     libraryDependencies ++= Seq(
       log4catsNoop.value
     ),
-    useYarn := true,
-    yarnExtraArgs += "--frozen-lockfile",
     jsVersionIntroduced("0.23.7"),
   )
   .dependsOn(
@@ -674,13 +672,10 @@ lazy val emberClient = libraryCrossProject("ember-client")
       log4catsSlf4j
     )
   )
-  .jsEnablePlugins(ScalaJSBundlerPlugin)
   .jsSettings(
     libraryDependencies ++= Seq(
       log4catsNoop.value
     ),
-    useYarn := true,
-    yarnExtraArgs += "--frozen-lockfile",
     jsVersionIntroduced("0.23.5"),
   )
   .dependsOn(emberCore % "compile;test->test", client % "compile;test->test")
@@ -1356,10 +1351,8 @@ def exampleProject(name: String) =
 def exampleJSProject(name: String) =
   http4sProject(name)
     .in(file(name.replace("examples-", "examples/")))
-    .enablePlugins(NoPublishPlugin, ScalaJSBundlerPlugin)
+    .enablePlugins(ScalaJSPlugin, NoPublishPlugin)
     .settings(
-      useYarn := true,
-      yarnExtraArgs += "--frozen-lockfile",
       scalaJSUseMainModuleInitializer := true,
       scalaJSLinkerConfig ~= (_.withModuleKind(ModuleKind.CommonJSModule)),
     )
