--- ember-core/js/src/main/scala/org/http4s/ember/core/h2/facade/Decompressor.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2019 http4s.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.http4s.ember.core.h2.facade
-
-import scala.annotation.nowarn
-import scala.scalajs.js
-
-@js.native
-@nowarn
-private[h2] trait Decompressor extends js.Object {
-  def write(raw: js.typedarray.Uint8Array): Unit = js.native
-  def execute(): Unit = js.native
-  def read(): Header = js.native
-}
-
-private[h2] object Decompressor {
-  def apply(options: HpackOptions): Decompressor =
-    hpackjs.decompressor.create(options).asInstanceOf[Decompressor]
-}
